requirements_pip.txt
pandas==2.2.3
pyyaml==6.0.2
requests==2.32.3